tests(TextArea)
